requirements.txt
# File Name: requirements.txt 
# Version: 1.2
# Last Updated: 20-Feb-2025
# Description: This file lists all the dependencies required for the application to run

# Flask is the lightweight web server framework for the application
Flask==2.3.3

# MySQL Connector for Python to enable database connectivity
mysql-connector-python==8.1.0

# SQLAlchemy for Object Relational Mapping (ORM) to manage database models
SQLAlchemy==1.4.41

# Werkzeug is required for password hashing and general security utilities
Werkzeug==2.3.7

# python-dotenv to manage environment variables securely
python-dotenv==1.0.0

# Optional: MarkupSafe is a Flask dependency, included explicitly to avoid version conflicts
MarkupSafe==2.1.3

# Flask-SQLAlchemy for SQLAlchemy support in Flask applications
Flask-SQLAlchemy==2.5.1

# Flask-Login for user session management
Flask-Login==0.6.2